refactor: split codex live and session usage helpers
